Add iids to milestones. Moved iids logic to separate concern
