+++ app/models/concerns/issuable.rb
@@ -16,8 +16,6 @@ module Issuable
 
     validates :author, presence: true
     validates :title, presence: true, length: { within: 0..255 }
-    validate :set_iid, on: :create
-    validates :iid, presence: true, numericality: true
 
     scope :authored, ->(user) { where(author_id: user) }
     scope :assigned_to, ->(u) { where(assignee_id: u.id)}
@@ -45,15 +43,6 @@ module Issuable
     def search(query)
       where("title like :query", query: "%#{query}%")
     end
-  end
-
-  def set_iid
-    max_iid = project.send(self.class.name.tableize).maximum(:iid)
-    self.iid = max_iid.to_i + 1
-  end
-
-  def to_param
-    iid.to_s
   end
 
   def today?
