+++ db/migrate/20130821090531_add_internal_ids_to_milestones.rb
@@ -1,0 +1,5 @@
+class AddInternalIdsToMilestones < ActiveRecord::Migration
+  def change
+    add_column :milestones, :iid, :integer
+  end
+end
